+++ packages/server/graphql/private/mutations/loginSAML.ts
@@ -2,6 +2,8 @@ import base64url from 'base64url'
 import getSSODomainFromEmail from 'parabol-client/utils/getSSODomainFromEmail'
 import querystring from 'querystring'
 import * as samlify from 'samlify'
+import {InvoiceItemType} from '../../../../client/types/constEnums'
+import adjustUserCount from '../../../billing/helpers/adjustUserCount'
 import AuthToken from '../../../database/types/AuthToken'
 import User from '../../../database/types/User'
 import generateUID from '../../../generateUID'
@@ -12,6 +14,7 @@ import encodeAuthToken from '../../../utils/encodeAuthToken'
 import {isSingleTenantSSO} from '../../../utils/getSAMLURLFromEmail'
 import {getSSOMetadataFromURL} from '../../../utils/getSSOMetadataFromURL'
 import {samlXMLValidator} from '../../../utils/samlXMLValidator'
+import sendToSentry from '../../../utils/sendToSentry'
 import standardError from '../../../utils/standardError'
 import bootstrapNewUser from '../../mutations/helpers/bootstrapNewUser'
 import getSignOnURL from '../../public/mutations/helpers/SAMLHelpers/getSignOnURL'
@@ -24,7 +27,8 @@ samlify.setSchemaValidator(samlXMLValidator)
 
 const CLAIM_SPEC = {
   'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress': 'email',
-  'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name': 'displayname'
+  'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name': 'name',
+  'http://schemas.microsoft.com/identity/claims/displayname': 'displayname'
 }
 
 const getRelayState = (body: querystring.ParsedUrlQuery) => {
@@ -47,7 +51,7 @@ const loginSAML: MutationResolvers['loginSAML'] = async (
   const normalizedName = samlName.trim().toLowerCase()
   const body = querystring.parse(queryString)
   const relayState = getRelayState(body)
-  const {isInvited, metadataURL: newMetadataURL} = relayState
+  const {metadataURL: newMetadataURL, isInvited} = relayState
   const doc = await dataLoader.get('saml').load(normalizedName)
   dataLoader.get('saml').clear(normalizedName)
 
@@ -57,7 +61,7 @@ const loginSAML: MutationResolvers['loginSAML'] = async (
         message: `Ask customer service to enable SSO for ${normalizedName}.`
       }
     }
-  const {domains, metadata: existingMetadata} = doc
+  const {domains, metadata: existingMetadata, orgId, samlOrgAttribute} = doc
   const newMetadata = newMetadataURL ? await getSSOMetadataFromURL(newMetadataURL) : undefined
   if (newMetadata instanceof Error) {
     return standardError(newMetadata)
@@ -82,7 +86,7 @@ const loginSAML: MutationResolvers['loginSAML'] = async (
   }
 
   const {extract} = loginResponse
-  const {attributes, nameID: name} = extract
+  const {attributes, nameID} = extract
   const normalizedAttributes = Object.fromEntries(
     Object.entries(attributes).map(([key, value]) => {
       const normalizedKey = CLAIM_SPEC[key as keyof typeof CLAIM_SPEC] ?? key.toLowerCase()
@@ -93,8 +97,9 @@ const loginSAML: MutationResolvers['loginSAML'] = async (
       return [normalizedKey, String(value)]
     })
   )
-  const {email: inputEmail, emailaddress, displayname} = normalizedAttributes
-  const preferredName = displayname || name
+
+  const {email: inputEmail, emailaddress, displayname, name} = normalizedAttributes
+  const preferredName = displayname || name || nameID
   const email = inputEmail?.toLowerCase() || emailaddress?.toLowerCase()
   if (!email) {
     return standardError(
@@ -151,7 +156,47 @@ const loginSAML: MutationResolvers['loginSAML'] = async (
     tier: 'enterprise'
   })
 
-  const authToken = await bootstrapNewUser(tempUser, !isInvited, dataLoader)
+  // find orgs specified in the SAML claim
+  const orgIds = await (async () => {
+    if (!samlOrgAttribute) {
+      return []
+    }
+    const samlOrgIdsAttribute = attributes[samlOrgAttribute]
+    if (!samlOrgIdsAttribute) {
+      // This is probably a misconfiguration but should not stop us from accepting the new user
+      sendToSentry(
+        new Error(
+          `The SAML attribute ${samlOrgAttribute} is missing from the SAML response. The following attributes were included: ${Object.keys(attributes).join(', ')}`
+        )
+      )
+      return []
+    }
+    const samlOrgIds = Array.isArray(samlOrgIdsAttribute)
+      ? samlOrgIdsAttribute
+      : [samlOrgIdsAttribute]
+    if (samlOrgIds.length === 0) {
+      return []
+    }
+    const orgs = await pg
+      .selectFrom('Organization')
+      .select('id')
+      .where('samlId', 'in', samlOrgIds)
+      .execute()
+    return orgs.map(({id}) => id)
+  })()
+
+  // if we don't provision via SAML claim, we default to the orgId from the SAML document
+  if (!samlOrgAttribute && orgId) {
+    orgIds.push(orgId)
+  }
+
+  const isOrganic = orgIds.length > 0 && !isInvited
+  const authToken = await bootstrapNewUser(tempUser, !isOrganic, dataLoader)
+
+  await Promise.all(
+    orgIds.map((orgId) => adjustUserCount(userId, orgId, InvoiceItemType.ADD_USER, dataLoader))
+  )
+
   return {
     userId,
     authToken: encodeAuthToken(authToken),