+++ packages/server/postgres/migrations/2025-03-25T13:33:26.204Z_addSAMLOrganizationAttributes.ts
@@ -1,0 +1,11 @@
+import type {Kysely} from 'kysely'
+
+export async function up(db: Kysely<any>): Promise<void> {
+  db.schema.alterTable('Organization').addColumn('samlId', 'varchar(100)').execute()
+  db.schema.alterTable('SAML').addColumn('samlOrgAttribute', 'varchar(100)').execute()
+}
+
+export async function down(db: Kysely<any>): Promise<void> {
+  db.schema.alterTable('Organization').dropColumn('samlId').execute()
+  db.schema.alterTable('SAML').dropColumn('samlOrgAttribute').execute()
+}